Fix test_retransmission_with_python_receiver
Three fixes to get the retransmission integration test passing:

1. retransmission_integration.rs: Send SETRATEANCHORTIME before streaming
   The test uses PcmStreamer directly (bypassing AirPlayClient), which
   never sends SETRATEANCHORTIME. Without it, AudioBuffered.play() blocks
   on rtspconn.poll(None) waiting for the "play" command, so 0 bytes were
   written to the audio output file.

2. audio.py: Move fini_audio_sink() to Audio base class and call it from
   AudioBuffered.serve()'s finally block, so the audio sink is properly
   closed after buffered audio streaming ends (not just realtime).

3. python_receiver.rs: Fix i16 overflow panic in verify_sine_wave_quality
   Cast to i32 before calling .abs() to avoid panic when min_sample is
   i16::MIN (-32768), whose abs() overflows i16 in debug builds.

+++ integration_tests/tests/common/python_receiver.rs
@@ -433,7 +433,8 @@ impl ReceiverOutput {
         }
 
         // 2. Verify dynamic range (should use most of 16-bit range)
-        if max_sample.abs() < 25000 || min_sample.abs() < 25000 {
+        // Cast to i32 before abs() to avoid overflow when min_sample == i16::MIN (-32768)
+        if (max_sample as i32).abs() < 25000 || (min_sample as i32).abs() < 25000 {
             tracing::warn!(
                 "Audio not using full dynamic range: max={}, min={}",
                 max_sample,
